first try lab2-checkpoint 1

+++ src/type/decimal_type.cpp
@@ -271,7 +271,7 @@ Value DecimalType::CastAs(const Value &val, const TypeId type_id) const {
       if (val.IsNull()) {
         return Value(type_id, BUSTUB_INT8_NULL);
       }
-      if (val.GetAs<double>() > BUSTUB_INT8_MAX || val.GetAs<double>() < BUSTUB_INT8_MIN) {
+      if (val.GetAs<double>() > BUSTUB_INT8_MAX || val.GetAs<double>() < static_cast<double>(BUSTUB_INT8_MIN)) {
         throw Exception(ExceptionType::OUT_OF_RANGE, "Numeric value out of range.");
       }
       return Value(type_id, static_cast<int8_t>(val.GetAs<double>()));
@@ -289,7 +289,7 @@ Value DecimalType::CastAs(const Value &val, const TypeId type_id) const {
       if (val.IsNull()) {
         return Value(type_id, BUSTUB_INT32_NULL);
       }
-      if (val.GetAs<double>() > BUSTUB_INT32_MAX || val.GetAs<double>() < BUSTUB_INT32_MIN) {
+      if (val.GetAs<double>() > BUSTUB_INT32_MAX || val.GetAs<double>() < static_cast<double>(BUSTUB_INT32_MIN)) {
         throw Exception(ExceptionType::OUT_OF_RANGE, "Numeric value out of range.");
       }
       return Value(type_id, static_cast<int32_t>(val.GetAs<double>()));
@@ -298,7 +298,8 @@ Value DecimalType::CastAs(const Value &val, const TypeId type_id) const {
       if (val.IsNull()) {
         return Value(type_id, BUSTUB_INT64_NULL);
       }
-      if (val.GetAs<double>() > BUSTUB_INT64_MAX || val.GetAs<double>() < BUSTUB_INT64_MIN) {
+      if (val.GetAs<double>() > static_cast<double>(BUSTUB_INT64_MAX) ||
+          val.GetAs<double>() < static_cast<double>(BUSTUB_INT64_MIN)) {
         throw Exception(ExceptionType::OUT_OF_RANGE, "Numeric value out of range.");
       }
       return Value(type_id, static_cast<int64_t>(val.GetAs<double>()));
